+++ src/main/java/org/folio/search/model/types/ResourceType.java
@@ -39,4 +39,8 @@ public enum ResourceType {
     }
     return UNKNOWN;
   }
+
+  public boolean isLinkedDataResource() {
+    return name.startsWith("linked-data");
+  }
 }