+++ src/test/java/org/folio/search/support/base/BaseConsortiumIntegrationTest.java
@@ -1,28 +1,18 @@
 package org.folio.search.support.base;
 
-import static java.util.Arrays.asList;
-import static java.util.Collections.emptyList;
 import static org.folio.search.support.base.ApiEndpoints.instanceSearchPath;
 import static org.folio.search.utils.TestConstants.CENTRAL_TENANT_ID;
 import static org.folio.search.utils.TestConstants.MEMBER_TENANT_ID;
 import static org.folio.search.utils.TestConstants.TENANT_ID;
 import static org.folio.search.utils.TestUtils.asJsonString;
-import static org.folio.search.utils.TestUtils.randomId;
-import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import java.util.List;
 import java.util.Map;
-import java.util.function.Consumer;
 import lombok.SneakyThrows;
-import org.folio.search.domain.dto.Instance;
-import org.folio.tenant.domain.dto.Parameter;
-import org.folio.tenant.domain.dto.TenantAttributes;
 import org.junit.jupiter.api.AfterAll;
 import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.ResultMatcher;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 
 public abstract class BaseConsortiumIntegrationTest extends BaseIntegrationTest {
@@ -31,38 +21,6 @@ public abstract class BaseConsortiumIntegrationTest extends BaseIntegrationTest 
   protected static void removeTenant() {
     removeTenant(TENANT_ID);
     removeTenant(CENTRAL_TENANT_ID);
-  }
-
-  @SneakyThrows
-  protected static void setUpTenant(String tenantName, Instance... instances) {
-    setUpTenant(tenantName, emptyList(), instances);
-  }
-
-  @SneakyThrows
-  protected static void setUpTenant(String tenantName, List<ResultMatcher> matchers, Instance... instances) {
-    setUpTenant(tenantName, instanceSearchPath(), () -> { }, asList(instances), instances.length, matchers,
-      instance -> inventoryApi.createInstance(tenantName, instance));
-  }
-
-  @SneakyThrows
-  protected static <T> void setUpTenant(String tenant, String validationPath, Runnable postInitAction,
-                                        List<T> records, Integer expectedCount, List<ResultMatcher> matchers,
-                                        Consumer<T> consumer) {
-    enableTenant(tenant);
-    postInitAction.run();
-    saveRecords(tenant, validationPath, records, expectedCount, matchers, consumer);
-  }
-
-  @SneakyThrows
-  protected static void enableTenant(String tenant) {
-    var tenantAttributes = new TenantAttributes().moduleTo("mod-search");
-    tenantAttributes.addParametersItem(new Parameter("centralTenantId").value(CENTRAL_TENANT_ID));
-
-    mockMvc.perform(post("/_/tenant", randomId())
-        .content(asJsonString(tenantAttributes))
-        .headers(defaultHeaders(tenant))
-        .contentType(APPLICATION_JSON))
-      .andExpect(status().isNoContent());
   }
 
   @SneakyThrows